+++ scripts/build-styles.mjs
@@ -82,6 +82,7 @@ await writeFile(
     '@import "../../src/components/link/link.css";',
     '@import "../../src/components/list/list.css";',
     '@import "../../src/components/table/table.css";',
+    '@import "../../src/components/toolbar/toolbar.css";',
     '@import "../../src/components/stack/stack.css";',
     '@import "../../src/components/grid/grid.css";',
     '@import "../../src/components/container/container.css";',